+++ src/lang_utils/error_codes.ml
@@ -181,7 +181,7 @@ let error_codes : (string * string option) list =
     "M0193", None; (* Can't declare actor class to have `async*` result *)
     "M0196", None; (* `system` capability supplied but not required *)
     "M0197", Some([%blob "lang_utils/error_codes/M0197.md"]); (* `system` capability required *)
-    "M0199", Some([%blob "lang_utils/error_codes/M0199.md"]); (* Deprecated experimental stable memory *)
+    "M0199", None; (* Deprecated experimental stable memory *)
     "M0200", Some([%blob "lang_utils/error_codes/M0200.md"]); (* Cannot determine subtyping or equality *)
     "M0201", None; (* Migration produces/consumes non-stable object *)
     "M0202", None; (* Migration produces/consume non-object type *)